Fix checkstyle plugin to run again (#46)
The checkstyle configuration within the pluginManagement section of the
pom was not being used when running the plugin. To get the configuration
to be used, the plugin configuration was removed from pluginManagement
and moved directly to the plugin within the build section.

The checkstyle jar version was also upgraded to 8.1. 8.2 is the latest
version, but it seems to have a bug where it blows up on escaped html
tags like '&lt;' or '&gt;' within javadoc.