Adding integration test module, not working
This module contains a generated service, api and client. I have started to add a reverse proxy integration test but haven't figure it out. Currently, if you run the docker images seperately, you can hit the service from the reverse proxy but when you try to make request over ssl, it fails. Not sure why, it should be routing correctly with the names